+++ src/assets/js/router.js
@@ -1,6 +1,5 @@
 import Vue from 'vue'
 import Router from 'vue-router'
-import Index from '../../views/Index'
 
 Vue.use(Router);
 
@@ -11,20 +10,20 @@ export default new Router({
         {
             path: '/:key(0{0}|[a-zA-Z0-9]{3,8})',
             name: 'index',
-            component: Index
+            component: () => import(/* webpackChunkName: "index" */ '../../views/Index')
         },
         {
             path: '/What_are_you_nong_sha_lei?',
             name: 'NotFound',
-            component: () => import(/* webpackChunkName: "not_found" */ '../../views/NotFound.vue')
+            component: () => import(/* webpackChunkName: "not_found" */ '../../views/NotFound')
         },
         {
             path: '*',
             redirect: '/What_are_you_nong_sha_lei?'
         },
         {
-            path: 'admin',
-            name: 'admin',
+            path: 'pasteme-admin',
+            name: 'pasteme-admin',
             // TODO
         },
     ]