+++ src/components/Footer.vue
@@ -2,10 +2,19 @@
     <div class="row">
         <div class="col-md-12">
             <div class="footer">
-                <p><a id="one-word" style="cursor: pointer;" @click="refresh">{{ oneWord }}</a></p>
+                <p><a id="one-word" @click="refresh">{{ oneWord }}</a></p>
+                <b-popover
+                        target="one-word"
+                        triggers="hover"
+                        placement="top">
+                    <a id="one-popover">{{ 0 >= cut_down_time ? $t('lang.footer.tooltip.refresh') :
+                        $t('lang.footer.tooltip.wait', { sec: cut_down_time }) }}</a>
+                </b-popover>
                 <p>
                     <a href='http://blog.lucien.ink' target='_blank'>Lucien's Blog</a>
-                    <a v-for="footer in $store.state.config.footer" v-bind:key="footer.id">&nbsp;&nbsp;|&nbsp;&nbsp;<a :href="footer.url" target="_blank">{{ footer.text }}</a></a>
+                    <a v-for="footer in $store.state.config.footer"
+                       v-bind:key="footer.id">&nbsp;&nbsp;|&nbsp;&nbsp;<a
+                            :href="footer.url" target="_blank">{{ footer.text }}</a></a>
                 </p>
                 <p>
                     <a>&copy; 2018 - {{ year }} </a>
@@ -14,7 +23,6 @@
                 </p>
             </div>
         </div>
-        <b-tooltip target="one-word" placement="topright">{{ $t('lang.footer.tooltip') }}</b-tooltip>
     </div>
 </template>
 
@@ -25,27 +33,42 @@
             return {
                 oneWord: 'Loading...',
                 year: new Date().getFullYear(),
+                cut_down_time: 0,
             }
         },
         mounted() {
-            this.refresh();
+            this.getOne().then(result => {
+                this.oneWord = result;
+            })
         },
         methods: {
-            async getOne() {
-                let one = null;
-                do {
-                    await this.api.get('https://v1.hitokoto.cn', {
-                        encode: 'text'
-                    }).then(response => {
-                        one = response;
-                    });
-                } while (one.replace(/[\u4e00-\u9fa5]/ig, '**').length > 100);
-                return one;
+            getOne() {
+                return this.api.get('https://v1.hitokoto.cn', {
+                    encode: 'text'
+                });
             },
             refresh() {
-                this.getOne().then(result => {
-                    this.oneWord = result;
-                });
+                if (this.cut_down_time === 0) {
+                    this.cut_down_time = -1;
+                    this.oneWord = 'Loading...';
+                    this.getOne().then(result => {
+                        this.oneWord = result;
+                        this.cut_down_time = 5;
+                        let clock = window.setInterval(() => {
+                            --this.cut_down_time;
+                            if (this.cut_down_time === 0) {
+                                window.clearInterval(clock);
+                            }
+                        }, 1000);
+                    });
+                }
+            },
+            makeToast(title, message, append = false) {
+                this.$bvToast.toast(message, {
+                    title: title,
+                    autoHideDelay: 500,
+                    appendToast: append
+                })
             }
         }
     }
@@ -61,6 +84,7 @@
     .footer a:link, .footer a:visited {
         color: #38488f;
     }
+
     #one-word {
         -webkit-touch-callout: none;
         -webkit-user-select: none;
@@ -68,5 +92,10 @@
         -moz-user-select: none;
         -ms-user-select: none;
         user-select: none;
+        cursor: pointer;
+    }
+
+    #one-popover {
+        font-family: Menlo, Monaco, "Andale Mono", "lucida console", "Courier New", monospace;
     }
 </style>