+++ src/views/Index.vue
@@ -1,27 +1,27 @@
 <template>
     <transition name="component-fade" mode="out-in">
-        <component :is="view"></component>
+        <component :is="view" v-bind="$data"></component>
     </transition>
 </template>
 
 <script>
-    import Form from '../components/Form'
-    import Success from '../components/Success'
-    import PasswordAuth from '../components/PasswordAuth'
-    import PasteView from '../components/PasteView'
-    import Loading from '../components/Loading'
+    import stateMixins from "../assets/js/mixins/stateMixin";
+    import {mapGetters} from "vuex"
     export default {
         name: "Index",
+        mixins: [stateMixins],
         data() {
-            return {
-                view: 'loading',
-                lang: null,
-                content: null,
-                placeholder: null,
-            }
+            return {}
+        },
+        computed: {
+            ...mapGetters([
+                "view",
+                "lang",
+                "content"
+            ])
         },
         watch: {
-            '$route.params.key': function () {
+            "$route.params.key": function () {
                 this.init();
             }
         },
@@ -30,38 +30,41 @@
         },
         methods: {
             init() {
-                this.$store.commit('init');
-                if (this.$route.params.key === '') {
-                    this.view = 'home';
+                this.$store.commit("init");
+                if (this.$route.params.key === "") {
+                    this.updateView("home");
                 } else {
-                    this.view = 'loading';
-                    this.api.get(this.$store.state.config.api + this.$route.params.key, {
+                    this.updateView("loading");
+                    this.api.get(this.$store.getters.config.api + this.$route.params.key, {
                         json: true
                     }).then(response => {
                         if (response.status === 200) {
-                            this.view = 'paste_view';
-                            this.content = response.content;
-                            this.lang = response.lang === 'plain' ? 'plaintext' : response.lang;
+                            this.updateView("paste_view");
+                            this.updateContent(response.content);
+                            this.updateLang(response.lang === "plain" ? "plaintext" : response.lang);
                         } else if (response.status === 401) {
-                            this.view = 'password_auth';
-                        } else if (response.status === 404 && this.$route.params.key.search('[a-zA-Z]{1}') !== -1) {
-                            this.$store.commit('updateMode', {
+                            this.updateView("password_auth");
+                        } else if (response.status === 403) {
+                            this.updateView("manual_deleted");
+                        } else if (response.status === 404 && this.$route.params.key.search("[a-zA-Z]{1}") !== -1) {
+                            this.$store.commit("updateMode", {
                                 read_once: true,
                             });
-                            this.view = 'home';
+                            this.updateView("home");
                         } else {
-                            this.$router.push('What_are_you_nong_sha_lei?');
+                            this.$router.push("What_are_you_nong_sha_lei?");
                         }
                     });
                 }
             },
         },
         components: {
-            'home': Form,
-            'success': Success,
-            'password_auth': PasswordAuth,
-            'paste_view': PasteView,
-            'loading': Loading,
+            "home": () => import(/* webpackChunkName: "home" */ "../components/Form"),
+            "success": () => import(/* webpackChunkName: "success" */ "../components/Success"),
+            "password_auth": () => import(/* webpackChunkName: "password_auth" */ "../components/PasswordAuth"),
+            "paste_view": () => import(/* webpackChunkName: "paste_view" */ "../components/PasteView"),
+            "loading": () => import(/* webpackChunkName: "loading" */ "../components/Loading"),
+            "manual_deleted": () => import(/* webpackChunkName: "manual_deleted" */ "../components/ManualDeleted")
         }
     }
 </script>